+++ HalgarisRPGLoot/Analyzers/ObjectEffectsAnalyzer.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using HalgarisRPGLoot.DataModels;
+using HalgarisRPGLoot.Settings;
 using Mutagen.Bethesda;
 using Mutagen.Bethesda.Plugins;
 using Mutagen.Bethesda.Skyrim;
